refactor(work): move /work runs onto pg-boss worker
/work was the last AI path running inline in Next.js. Refactored
so every backend.run() lives in the worker process — same
architectural rule the dashboard chat path already follows.

Wire shape:

- Browser POSTs /api/work/threads/{threadId}/runs.
- Route inserts work_run + processing_job, enqueues WORK_RUN to
  pg-boss, returns {runId, backend}. Atomic-ish: any failure in
  the insert+enqueue path finalizes both rows so the UI never
  spins forever.
- Worker picks up WORK_RUN, runs the agent, writes each AgentEvent
  into work_run_event by seq.
- Browser opens an EventSource on
  /api/work/threads/{threadId}/runs/{runId}/events. The route
  long-polls work_run_event by seq; each frame is tagged with
  `id: <seq>` so EventSource auto-reconnect resumes via
  Last-Event-ID without replaying. Closes when work_run reaches
  a terminal status, OR after MAX_LIFETIME_MS so a long-running
  pg client doesn't hold idle.

Files moved:

- packages/llm/src/work/store.ts: canonical work_* table helpers.
  Web's @/lib/work-store now re-exports from here so the worker
  can use the same code (it can't import from apps/web).
- apps/web/src/lib/work-store.ts: re-export shim.

New:

- packages/db/src/jobs.ts: WORK_RUN queue name + WorkRunPayload.
- apps/worker/src/jobs/work-run.ts: ports the inline POST flow
  exactly — same prompt build, same mcpServers wiring, same
  auto-memory pipeline kick. The only thing dropped is the inline
  SSE emit (worker writes to DB; the events route does the SSE).
- apps/web/src/app/api/work/threads/[threadId]/runs/[runId]/events/
  route.ts: SSE long-poll tail with Last-Event-ID resumption.

Modified:

- apps/worker/src/index.ts: register WORK_RUN handler with N
  workers (matches metric_refresh shape so a slow Hermes spawn for
  one thread doesn't HOL-block another).
- apps/web/src/app/api/work/threads/[threadId]/runs/route.ts:
  gutted from 237 lines of inline ReadableStream to a thin
  enqueue-and-return. Atomic-ish rollback on enqueue failure.
- apps/web/src/app/work/work-screen.tsx: switched from
  res.body.getReader() SSE parsing to a real EventSource on the
  events endpoint.
- apps/worker/src/reconciler.ts: minAgeMs<=0 skips the time
  predicate (race between Postgres now() and JS Date.now() at
  sub-ms resolution was making the orphan reconcile test flaky).

Known degradation, follow-up:

- /api/work/runs/{runId}/cancel was using an in-process
  AbortController registry on the web side. Now that runs execute
  in the worker, that registry is empty and cancel silently
  no-ops. Restoring cancel needs a sentinel in work_run (or a
  shared table) that the worker handler polls and aborts on. Not
  blocking the architectural correction.

All 283 tests pass.
